+++ src/models.rs
@@ -9,8 +9,8 @@ pub mod educator_model;
 #[derive(Debug, Deserialize, Serialize)]
 pub struct User {
     pub name: String,
-    pub watch_educators: Vec<u32>,
-    pub watch_groups: Vec<u32>,
+    pub watch_educators: Vec<i64>,
+    pub watch_groups: Vec<i64>,
     pub email: String,
 }
